third_party/WebKit/LayoutTests/http/tests/media/media-source/webkitmediasource-add-and-remove-buffers-expected.txt

CONSOLE WARNING: 'WebKitMediaSource' is deprecated. Please use 'MediaSource' instead.
Tests webkitSourceAddId() & webkitSourceRemoveId() methods

Test adding an ID while closed.
Got expected exception InvalidStateError: Failed to execute 'addSourceBuffer' on 'WebKitMediaSource': This MediaSource's readyState is not 'open'.

running testAddBufferFailureCases
EVENT(webkitsourceopen)
Test empty type.
Got expected exception InvalidAccessError: Failed to execute 'addSourceBuffer' on 'WebKitMediaSource': The type provided is empty.
Test an unsupported type.
Got expected exception NotSupportedError: Failed to execute 'addSourceBuffer' on 'WebKitMediaSource': The type provided ('audio/x-unsupported-format') is not supported.
Test a supported type with an unsupported codec.
Got expected exception NotSupportedError: Failed to execute 'addSourceBuffer' on 'WebKitMediaSource': The type provided ('video/webm; codecs="vp8, speex"') is not supported.
Test reaching sourceID limit.
Test that SourceBuffers can't be added while in the ended state.
Got expected exception InvalidStateError: Failed to execute 'addSourceBuffer' on 'WebKitMediaSource': This MediaSource's readyState is not 'open'.

running testRemoveNullBuffer
EVENT(webkitsourceopen)
Test null buffer case
Got expected exception InvalidAccessError: Failed to execute 'removeSourceBuffer' on 'WebKitMediaSource': The SourceBuffer provided is invalid.

running testRemoveAgain
EVENT(webkitsourceopen)
Test removing a buffer that was already removed.
Got expected exception InvalidStateError: Failed to execute 'removeSourceBuffer' on 'WebKitMediaSource': This MediaSource does not contain any SourceBuffers.

running testRemoveBufferAfterEnded
EVENT(webkitsourceopen)
Test that a buffer can be removed while in the ended state.

running testAddBufferAfterRemoving
EVENT(webkitsourceopen)
Test that a buffer can be added again after it is removed.

running testAppendFailureCases
EVENT(webkitsourceopen)
Test a successful append.
Test append with a null buffer.
Got expected exception InvalidAccessError: Failed to execute 'append' on 'WebKitSourceBuffer': The data array provided is invalid.
Test append after buffer has been removed.
Got expected exception InvalidStateError: Failed to execute 'append' on 'WebKitSourceBuffer': This SourceBuffer has been removed from its parent MediaSource.
END OF TEST
